JIT: consistently mark method table accesses as invariant (#45854)
Introduce a new utility method to create the IR for method table access,
marking the resulting indir as invariant. This allows method table access
CSEs (which will be increasingly common with the advent of PGO-enabled
Guarded Devirtualization).

Add a workaround to assertion prop for method table to be able to see
through a CSE'd method table access.

Remove a few optimizer inhibitions related to colon cond.