[wasm] provision-wasm: always install the "official" version (#45990)
if `EMSDK_PATH` was unset, or set to the default path, then
`provision-wasm` target would become a no-op. Instead, always install it
to the default path.

It can be useful to have the official version available, even if you
have a different one installed, so you can do comparisons etc. - Zoltan