[wasm][debugger][tests] Refactor test harness to support simpler test code (#44987)
* [wasm][debugger] Move DevToolsQueue to a separate file

* Move Inspector to separate file, mv Support.cs=>DebuggerTest.cs

* [wasm][debugger][tests] Some improvements to DevToolsQueue

- `Pump` -> `TryPump`
- use a ConcurrentQueue instead of a List, for tracking the pending
requests.

DevToolsClient:
- use DevToolsQueue, and remove the redundant code.
- Wake up the runloop when a new send task gets added

* [wasm][debugger] Check Proxy's client_initiated_close early

- this allows the proxy to shutdown quicker, and not have to wait for
the other tasks to get scheduled.

- Fixes the issue of chrome processes left around.

* [wasm][debugger][tests] Inspector - handle event being received before a waiter

.. gets added.

This handles the case where `NotifyOf` gets invoked before a
corresponding `WaitFor` gets called.

* [wasm][debugger][tests] Refactor test harness to support simpler test code

Problem:

- Currently, test methods must call `insp.Ready(..)` with a
callback. And `Ready` takes care of the init, and shutdown for
the proxy, and browser.

- But this means that we have to repeat that boiler plate code in all
the test methods. And the stack traces are deeply nested.

- Instead, we want to have the *test method* contain only the test
specific code.

Solution:

- we want to do that by using xunit's `IAsyncLifetime`, with
methods:
	- `InitializeAsync` - setup proxy+browser here
	- `DisposeAsync` - shutdown

- This would mean that when the *test method* runs, it already has
everything set up, so the code can be just specific to the actual test.

Implementation:

- Improve DevToolsClient's RunLoop, to support this
	- Tries to handle shutdown/close requests explicitly, and not throw
	errors in such cases

	- Adds a `RunLoopStopped` event, which gets invoked with the reason due
	to which it stopped. Eg - Exception getting thrown, cancellation,
	shutdown request etc.

	- With `insp.Ready(..)`, the whole setup (launch proxy, browser,
	run test code) happens from the same *test method*. But here we
	are doing it at separate times, from different methods, and it's
	all async.
		- IOW, there are multiple places where failure can
		happen, and we need to explicitly propogate that, and
		fail the appropriate listeners/tasks/etc
		-> causing the test to correctly fail.

Note: This does not change *any* actual test code! The `Ready` method is
adjusted to use the new init/shutdown bits.
Subsequent commits will make the test changes.

* [wasm][debugger][tests] Port some tests to the simpler test harness bits

We add a `SingleSessionTestBase` class, which implements
`IAsyncLifetime` to init/shutdown the proxy+browser outside the test
methods.

And port `HarnessTests`, and `SteppingTests` to that.

Added a few more tests to `HarnessTests`, to check for common cases like
test methods throwing exceptions, or timing out, or commands failing
etc.
	- This is needed because with the older `insp.Ready` API, the
	whole thing was called from the test method, so failures would
	correctly fail the test.

	- But now we do these things in separate steps, so failing the
	test requires a bit more work in the harness.

Note: I didn't port all the tests in one go, to make the review a bit
simpler. A subsequent PR can move rest of the tests over, and do some
cleanup that it would afford.

* [wasm][debugger][tests] Fix race in ExceptionThrownInJSOutOfBand test

Fix the race where the `Runtime.exceptionThrown` message can be received
before the test calls `WaitFor`, causing it to get missed.

* [wasm][debugger] Formatting fixes on the changes

* Remove unncessary AssertHelpers.ThrowsAsync

Suggested by @lewing !

* [wasm][debugger][tests] Fix a race when sending messages

DevToolsClient.Send adds to `pending_ops`, which can happen while the
list is being modified in `RunLoop`.

Fix: Since, we need to modify it outside `RunLoop` only when a new send
task is available, we can let `RunLoop` itself add it to the list. And use
a taskCompletionSource to wake up the loop.