[Wasm] Add AOT CI Leg (#46651)
In lieu of running all the library tests suites in AOT mode on CI (estimated to be ~36 hrs = 240 test suites * ~9min AOT time per test suite), only System.Buffers will be ran in AOT mode to standup the leg.  Once we find a more acceptable solution, the other libraries tests will be enabled. 

The CI leg will start out on the staging pipeline to avoid blocking and failing the build.  

Co-authored-by: Steve Pfister <steve.pfister@microsoft.com>
Co-authored-by: Mitchell Hwang <mitchell.hwang@microsoft.com>