[wasm] Add support for building out-of-tree projects, which use a local runtime build (#47555)
* [wasm] Add $( WasmBuildAppAfterThisTarget) to control which target

.. triggers WasmAppBuild. Defaults to `Publish`.

* [wasm] emcc, mono-aot-cross doesn't like relative output paths

eg. `-o foo/bar.js`

.. so convert that to absolute paths.

* [wasm] Check for WasmMainJSPath close to it's use

* [wasm] Add new $(WasmGenerateAppBundle) to allow skipping app bundle

.. generation. This is useful when you want to do all the other steps
for wasm, but handle the packaging/bundle generation yourself. Eg.
Blazor

* [wasm] Ensure that generated dotnet.{js,wasm} get copied to the app dir

This is usually done as part of the bundle generation (NativeAssets).
But in cases when that step is being skipped, we still want to copy over
generated dotnet.{js,wasm} (eg. in case of AOT) to the app dir. This is
the case that blazor uses.

* [wasm] Allow skipping mono-cil-strip step

In some cases, eg. blazor (till we fix it), `mono-cil-strip` can break
the app, because of bugs, or linker issues etc.

Also, do this only in the AOT case.

* [wasm] Add support for building wasm projects outside the tree

.. using local runtime builds.

* [wasm] fix up the support for local build dir

* [wasm] add comments for new properties

* [wasm] LocalBuild.*: add more comments, validation

* [wasm] sample: Use BeforeTargets for Prepare target, as a psuedo test for that