Set some pointers to null after free()ing them (#47415)
These free() calls are done in the middle of the method body; it's
possible that a duplicate free() can be called on these at the cleanup
code that runs before the method returns. So set these pointers to null
after free()ing them to avoid free() already free()d memory.