Improve error message when bundler fails due to duplicate entries (#48336)
There are multiple possible cases which will make the bundler fail with duplicate entries. Currently there's no indication of which entries caused the problem, making it really hard to diagnose these failures.

This change prints out the source file paths and the colliding relative bundle path in the error message. This should help with diagnosing the failures.